状態管理 storeファイルを分割
